Fix metadata

+++ src/yproxy.cpp
@@ -147,7 +147,7 @@ bool YProxyReader::read(char *buffer, size_t *amount) {
 
     auto rc = ::read(client_fd_, buffer, *amount);
     if (rc <= 0) {
-      elog(WARNING, "reacquiring connection on offset %d", current_chunk_offset_);
+      elog(WARNING, "reacquiring connection on offset %ld", current_chunk_offset_);
 
       if (++this->current_retry < this->retry_limit) {
         auto rrc = this->prepareYproxyConnection(order_[order_ptr_], current_chunk_offset_);
@@ -534,7 +534,7 @@ YProxyLister::message YProxyLister::readMessage() {
 
 std::vector<storageChunkMeta> YProxyLister::readObjectMetaBody(std::vector<char> *body) {
   std::vector<storageChunkMeta> res;
-  int i = PROTO_HEADER_SIZE;
+  size_t i = PROTO_HEADER_SIZE;
   while (i < body->size())
   {
     std::vector<char> buff;
